requirements.txt
opencv-python~=4.11.0.86
numpy~=2.1.3
scikit-image
scikit-learn~=1.6.1
matplotlib
pillow
tensorflow
torch
torchvision
